stockai/backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
yfinance>=0.2.36
pydantic>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
httpx>=0.25.0
